Improve Test for SON Bug 715 (V(*) and I(*) for .PRINT)
This commit improves the test for SON Bug 715, which tests V(*)
and I(*) on a .PRINT TRAN line.  The previous version of the .sh
file would detect if V(*) or I(*) emitted the wrong column names
in the output file.  However, it would not detect missing columns
that would occur if either V(*) or I(*) caused no output to occur.

SON Bug 715
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=715