Update to make this test more robust
This test occasionally fails in both serial and parallel.  From what I
have seen these are usually minor compare failures.

This is a sampling test that relies on 1000 Monte Carlo samples.  So,
the result it produces is statistical, rather than deterministic.  In
many of my sampling and embedded sampling tests, I use a lot more than
1000 samples, to make the computed output statistics more reliable.  I
also frequently will use a 2-pass loop, where if the test doesn't pass
the first time it is attempted, it is attempted a second time.

I've added the 2-pass loop to this test.  Hopefully that will help.  I
have kept the number of samples the same.