Add comparator tolerances to BSIMCMG test case
I discovered today, after months of not building Xyce with Intel
compilers on my Ubuntu laptop, that one test was suddenly failing that
has never failed anywhere else, and doesn't fail on the same machine
when Xyce is built with gcc.

The failure is for the "gummel_n_outvars" test case, which is meant to
test the implementation of gm, Vbs, Vdsat, etc. in the BSIMCMG version
100.  It tests GM by running to identical mosfets with very slightly
offset gate voltages, then doing finite difference to get dIDS/dgate,
which is then compared to the output gm.

The lack of an abstol on the finite-difference version of gm led to a
minor compare failure (1.47%) for one choice of gate voltage.  This
was so insignificant, though, the failure was only when this finite
difference was compared to the gold standard, not when compared to gm
itself.  So I'm adding an abstol to shut up the completely meaningless failure.