Fix false pass test issue
The test script would incorretly issue a pass for
the dev_doc_cat test if one file failed the diff
compairson but the next one passed.  This fixes
the logic by moving the initialization of the
return flag out of the loop.  Also the gold
standard data was out of date for the capacitor
as it did not have the charge parameter in the
tex file.