Removing an output for supernoding that has been eliminated
Supernoding was outputting a lot of information that is nice in serial,
but way too verbose for parallel.  Either way, the information is more
useful for developers than users.  In reducing the output for supernoding,
this output was removed, so it also needs to be removed from the test that
is looking for it.

SON Bug 1246