Removes output that does not print when test is run on >2 processors
Certain netlist information is not output in the same format when the
error is found on off-processors (processors that are not 0).  So, these
tests will fail for parallel regression testing using >2 processors.
This commit removes the error string checks that cannot be counted on for
parallel testing.  As well as increasing processors, this could also be
an issue for different distribution strategies.

SON Bug 1239