Test SON Bug 1238 for .NOISE Analysis
This commit adds a test for VR(*), VI(*), VM(*), VP(*), VDB(*),
IR(*), II(*), IM(*), IP(*) and IDB(*) for .PRINT NOISE.  Note that
only the branch currents for V and E devices are output for a
NOISE analysis. The lead currents for the R and C devices are not.

This commit also re-named the existing test bug1238, for .PRINT AC,
to be bug1238-ac.  That test is otherwise unchanged.

SON Bug 1238
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1238