Add .PRINT Wildcard Test for * in Node/Device Names
This commit adds a test for the use of * in node names and device
names.  Examples are node 1* and device R1*.  Those usages are not
recommended, but Xyce should do something sensible in those cases.

The Xyce behavior is now that, for example, .PRINT DC V(1*) is a request
for all nodal voltages that start with '1'.  It is not an explicit
request for the voltage at node 1*, if that node exists in the netlist.
The same now applies, for example, to .PRINT DC I(R1*) P(R1*) W(R1*).
R1* is now treated as a wildcard request for all devices that start
with "R1".

Changes were also made to several INVALID_CHARS tests to reflect this
change in Xyce behavior.

SON Bug 1211
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1211