Update gold Xyce netlits for two tests
The gold, hand translated accepted netlist, for two
XDM tests needed to be updated due to changes in
Xyce related to SON BUG 983.  In the past Xyce
would treat multiple V(*)'s on the print line as
one but the translated netlist ends up with
multiple V(*) because that's what is in the
original netlist.  This keeps the gold xyce
netlist consistent with what was requested.