Clean up an enormous mess
A merge of an ancient branch directly onto develop this week
completely hosed our public github update process for Xyce_Regression.

I have spent nearly the last 8 hours trying to rewrite history so that
I can recover things so that github can be updated, but after all that
work, the following files needed to be manually reconciled against
master in an ueberfix.  Consider it a conflict resolution commit.