Update Tests for .PRINT HOMOTOPY FORMAT=PROBE
This commit updates the tests for .PRINT HOMOTOPY FORMAT=PROBE. It
removes the tests for that now-removed feature in Output/HOMOTOPY
and Certification_Tests.  It adds a "probe format defaults to STD"
test as Output/HOMOTOPY/homotopy-probe-defaults-to-prn.  The final
change was to test that FORMAT=<TOUCHSTONE|TOUCHSTONE2> also
correctly defaults to FORMAT=STD for .PRINT HOMOTOPY.

SON Bug 939
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=939