Test <netlistName>.prn Output for MPDE Outputters
This commit adds testing of the <netlistName>.prn files produced
by the .PRINT TRAN lines in the netlists for these Output/MPDE tests.
One key change was that time integrator for MPDE was changed from
GEAR to TRAP.  That was done because the MPDE/WaMPDE interpolation
functions had not been implemented for the Gear method, which led
to only the DCOP values being output in those <netlistName>.prn
files when GEAR was used.

I also modified the mpde-prn.cir.sh file to make it more similar
to ones used for testing FORMAT=CSV and FORMAT=TECPLOT.

SON Bug 969
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=969