Hopefully Last Set of Changes to Output/MPDE Tests
This commit changed the "unsupported formats default to prn" tests
to use TRAP as the time integration method for MPDE.  I purposefully
left the MPDE time integration method as GEAR for the RAW format test.
That illustrates the bug with GEAR, but was sufficient to show that
the output was in RAW format.

SON Bug 969
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=969