Change Comparator for Two FORMAT=<CSV|NOINDEX> Tests
This commit changes the comparator from xyce_verify to file_compare
for the Output/DC/dc-csv and Output/TRAN/tran-prn-noindex tests.
That comparator works better for verifying the correctness of the
outputters for tests for table-based output that is not in
FORMAT=STD, when the output is "smooth" and can be output at a
known, fixed of time points for TRAN tests.

SON Bug 985
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=985