Add .sh File for Netlist That Uses FORMAT=NOINDEX
SON Bug 544 was about "Initial condition statement results in
failed simulation", because the test netlist would fail to run with
the development branch of Xyce 6.3.  The .PRINT TRAN line uses
FORMAT=NOINDEX, but does not include any variables on that .PRINT
line.

So, to avoid error messages from xyce_verify, this commit adds a
.sh file that simply checks that the simulation completes without
error, and that the <netlistName>.prn and <netlistName>.mt0 files
are made.  That testing approach meets the spirit of the bug,
without propagating an incorrect use of xyce_verify.pl with
FORMAT=NOINDEX on .PRINT line.

SON Bugs 544 and 985
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=544
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=985