Refactor FORMAT=CSV Tests for .HB Output
This commit refactors two FORMAT=CSV tests for .HB output.  In order
to use xyce_verify.pl to test the contents of the time-domain output
files, it converts the tests' time-domain .csv output to .prn output
(for verification purposes) with the convertToPrn2.py script.  It
also verifies that the FORMAT=CSV files, for time-domain output, do
not have a simulation footer.  This same approach was used to the
verify the time-domain output for the Output/MPDE tests.

SON Bug 985
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=985