Test Use of Default Device-Parameter Syntax for .PRINT
This commit adds a test for the use of the default device-parameter
syntax (e.g., R1 vs, R1:R) on a .PRINT line for the R, L, C and I
devices.  It covers the use of subcircuits.  It also verifies that
invalid parameters (e.g., R1:BLEEM or X1:R1:BLEEM) still produce
the correct parsing-error messages.

SON Bug 971
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=971