Add tests for BSIM-CMG 108
I have added BSIM-CMG 108 to the code and it needs testing.

The model is basically tested by the very same tests as BSIM-CMG 107,
with the addition of two tests to check the added multiplicity
feature.  The gold standards are just copied from 107, which
themselves were simply reformatting of output provided with the model
itself.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1258