Remove "p" from output/error file suffix
run_xyce_regressionMP was creating ".pstderr" and ".pstdout" files
instead of ".stderr" and ".stdout", presumably to aid debugging when
the script was being developed (to make them distinct from the serial
version).

Unfortunately, if this script were ever to be used instead of
run_xyce_regression in nightly testing, that would break how we report
problems to CDash.  I'm now removing the "p" from these suffixes.