Revert "Remove "p" from output/error file suffix"
This reverts commit 594d29d916eda14299f33d5eda0cdbb6a94abd48.

My assumption that simply changing the name of the output file would
have no ill effect was wrong, and now run_xyce_regressionMP spews some
output to the screen that it should not be spewing.

Since there is almost zero chance that this script will be getting
used in nightly regression anyway (it doesn't help on the one machine
where we really need help), being compatible with the way we report to
cdash is not important.  Reverting the commit to clean up the spew.