Add tests for BSIM-CMG model binning
Garrick Ng has added support to our binning code for devices that
should be binned on NFINMIN/NFINMAX instead of WMIN/WMAX.  That
basically means BSIM-CMG.

This commit adds tests of this binning capability for all three versions
of BSIM-CMG in Xyce.

They are patterned on the way I did bsim6 binning tests, but using a
copy of the gummel_n and gummel_p netlists that come with the model.
The binned library contains two model cards, one of them the one that
was used to generate the gold standard, and another one that is
sufficiently different that the comparison to gold standard would fail
badly if it were selected.

In all cases, the "correct" model card is the second in the file.

Gold standards are not simply archived copies of what the binned
circuit produces --- they're copies of the output of a simpler,
non-binned netlist that uses only the "correct" model card.  These are
in the BSIMCMG_GS_GENERATING directory, and will not be run by
run_xyce_regression (because the netlists don't have .cir suffixes).

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1200