Test Mixed Signal Methods to Get GIDs of Adj. Nodes
This commit adds partial tests for the getNumAdjNodesForDevice() and
getAdjGIDsForDevice() methods to the Python interface to the Mixed
Signal Interface.  The user can now request the GIDs (of voltage nodes
and the ground node) that are adjacent to a specified device.  The
return code and values for getNumAdjNodesForDevice() are verified
to be correct.  Only the return code is currently verified for the
getAdjGIDsForDevice() method.  That is because those GID values
may vary on different platforms.  So, I still need to account for
that in the .sh file for this new test.

SRN Bug 1052
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1052