Improve Mixed Signal Test for Getting GIDs of Adj. Nodes
This commit updates the test for the getNumAdjNodesForDevice() and
getAdjGIDsForDevice() methods to the Python interface to the Mixed
Signal Interface.  The .sh file now verifies that the returned
GID values are "consistent" based on the circuit topology, even
though those GID values may vary on different platforms.  So, the
details of the .sh file may need to be modified if the netlist
changes.

SRN Bug 1052
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1052