Initial Tests for FIND-WHEN and WHEN Measures for .AC
This commit adds the initial tests for the FIND-WHEN and FIND measures
for .AC analyses, for both the step and non-step cases. More tests
involving expressions, other measure values, and the comparision of two
voltage values in the WHEN clause still need to be added.  A test with
branch currents might also be added.

This commit also updates the tests for error messages emitted when an
supported measure type (e.g., ON_TYPE) is requested on a .MEASURE AC
or .MEASURE DC line.

SON Bug 1270
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1270