Test Expressions with FIND-WHEN Measures for .AC
This commit adds tests for the use of expressions with the FIND-WHEN
measure for .AC analyses, for both the step and non-step cases.  Note
that the the comparision of two voltage values in the WHEN clause of
these new tests only only matches HSPICE to within about 1%.  So, we
might revisit our interpolation algorithm for that case at some point.

SON Bug 1270
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1270