Test .PRINT AC line with .STEP and a .LIN Analysis
This commit adds tests for the use of .PRINT AC lines with .STEP
and .LIN.  The main purpose is to verify that the error checking
during the creation of the S(), Y() and Z() operators is now correct.
The format of the <netlistName>.s2p output file may need further
work though, for improved compliance with the Touchstone standard.

SON Bug 1271
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1271