Improve Testing for FIND-WHEN and WHEN Measures
This commit improves the testing of the FIND-WHEN and WHEN measures
for all three measure modes (AC, DC and TRAN).  In particular,
it adds/refines tests for the case of WHEN V(1)=V(2) where either
(or both) sides of the equality in the WHEN clause might be a varying
quantity.  This tests the improved interpolation algorithm that
was added to the IO::Measure::FindWhen class today.  The updated
gold standards for each test were verified against HSPICE results.

SON Bug 1273
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1273