Test Backward/Null Time Steps for Mixed Signal Interface
This commit tests the Xyce error messages for the simulateUntil()
method for the Mixed Signal Interface.  If the user requests the next
pause time (requestedUntilTime) to be less than or equal to the
current simulation time then a Xyce error message is emitted and
then the Xyce simulation aborts.  The Python script should exit
cleanly with a return code of 1.

SON Bug 1075
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1075