Tests for Valid Touchstone2 Input Files for YLIN Device
This commit adds the initial set of tests for valid Touchstone 2
input files for the YLIN device.  It uses a two-port network to
test the cases of:

  1) The input matrix being in "Full" format without any column wrap.

  2) The input Touchstone 2 file using Y-, Z- or S-parameters, in RI
format.

  3) The input Touchstone 2 file using Y-parameters, in MA and DB
format.

  4) All four frequency multipliers (Hz, kHz, MHz and GHz).

  5) All input files use [Two-Port Data Order] = 12_21.

The .HB analysis used to generate the gold standards, and the .LIN
analysis used to generate the various S-, Y- or Z-parameter
input files are also included in this commit.  Those netlists are
not actually run as part of these regression tests though.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215