Test Parsing of Short-Circuit Currents in YLIN Device
This commit adds a test for parsing a non-standard extension to
the Touchstone 2 file format for the YLIN device.  The input file
(<netlistName>.s2p) has additional columns in the [Network Data]
section of that file, for each frequency, for the per-port
short-circuit currents.  Those Isc values are not actually used
by the YLIN device yet, so this is just a parsing test.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215