Add a test for newly functional BSIM6 out vars
My work on bugs 410 and 788 have enabled automatic generation of some
Verilog-A output variables (those that are declared to have either a
"desc" or "units" attribute (Verilog-A standard), or which have the
"ask" attribute set to "yes" (pre-standard ADMS feature).

This commit tests a few of the BSIM6 output variables so generated.
It is patterened on a similar test for the BSIM-CMG 110, which
currently still has manually hacked-in output variables.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410