Add a test of PSP103 output variables
This is a copy of one of the PSP103 tests, hacked with two mosfets to
perform finite differencing on some output variables to compare to the
"gm" transconductance output variable.

This was a trickier one than some other MOSFET devices, because what
PSP calls "IDS" for an output variable is NOT the thing that they
differentiate to get "GM" --- it is only one term of the total
drain/source current.  The actual, total drain source current is the
internal variable "ids_i" that is NOT output, but which is
differentiated to get gm.

So I had to look at the code to figure out what sum of output
variables is actually equivalent to ids_i, and do finite differencing
on *that*.