Fix typoe in PSP102 outvars *COMP line
My various settings of tolerances on the one failing signal for PSP
seem to have been having no effect.  That's because there was a typoe
in the *COMP line, so none of the tolerances  were being applied (the
script does a straight-text match of the *COMP line and the .PRINT
line).