Test DERIV, INTEG and RMS Measures for .AC and .DC
This commit adds tests for the DERIV, INTEG and RMS measures for
.AC and .DC analyses.  It also updates the gold standards for the
RMS measure tests for .TRAN to reflect the new method for calculating
the RMS measure values.  Finally, the "gold stdout" for the AVG
measure tests for .AC and .DC were updated to reflect the corrected
FROM-TO measure window information, for the corner case of FROM=TO
for the AVG measure.

SON Bugs 1282 and 1283
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1282
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1283