Improve Testing for DERIV Measure for .TRAN
This commit updates the MEASURE/FindWhenVariableTest to verify
that the WHEN clause of a DERIV measure gets the same answer (time
value) as the equivalent WHEN measure for the case where those two
WHEN clauses use two voltages that are variable.  An example is
WHEN V(4)=V(2), where V(2) and V(4) are not constant.

SON Bugs 1273 and 1282
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1273
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1282