Test WEIGHT Qualifier for ERR1 and ERR2 Measures
This commit adds a subtest for the WEIGHT qualifier for the ERR1
and ERR2 measures.  Per the Reference Guide, "this parameter is not
implemented in Xyce, but is included for compatability with HSPICE
netlists."  So, the updated test in MEASURE_DC verifies that the
WEIGHT qualifier has no effect on the measure values, and does not
cause any parsing errors.

SON Bug 1278
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1278