Test .OPTIONS MEASURE and .OP for ERR1 and ERR2 Measures
This commit adds ERR1 measure statements to the existing tests for
the behavior of .MEASURE DC for .OP only, and for .OP  with .DC.  It
also tests .OPTIONS MEASURE DEFAULT_VAL and .OPTIONS MEASURE MEASFAIL
for the ERR1 and ERR2 measures for the TRAN measures.  Finally, it
removes one unused gold file.

SON Bug 1278
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1278