Initial Test Set for Continous Measures (FIND-WHEN)
This is the initial set of tests for "continuous" WHEN, FIND-WHEN
and FIND-AT measures for AC_CONT, DC_CONT and NOISE_CONT measure
modes.  They cover all valid RISE, FALL and CROSS values, except
for 0. The tests for TRAN_CONT mode are incomplete, and the
version for that mode in MEASURE_CONT/STEP is currently excluded.
The error message test (for measure types that are not supported
for "continuous measures") is only done for DC_CONT measures.

SON Bugs 1274 and 1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313