Change YLIN/opAmpWithIsc Test to use -norun
The YLIN/opAmpWithIsc test was intended to be a parsing test.
The ISC data in the input file was simply "made up", and will
cause convergence problems once the S-parameter model is updated
to actually use the short circuit currents in frequency domain.  So,
this commit changes the Xyce invocation to use -norun.  That approach
still verifies that the parsing code for ISC does not cause Xyce
to fail during parsing.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215