Test Parsing of Time-Domain Short-Circuit Currents in YLIN
This commit adds tests for parsing data for time-domain, per-port
short-circuit currents for the YLIN device.  Those Isc values are not
actually used by the YLIN device yet, so this is just a parsing test
that is run with Xyce -norun.  This commit also adds an errorexit
test for invalid values for the new ISC_TD_FILE and ISC_TD_FILE_FORMAT
model parameters.

SON Bug 1295
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1295