Test Upper and Lower Format for TS2 Input for YLIN
This commit adds the minimal (e.g. a 2-port) test of the [Network Data]
section of the Touchstone2 formatted input file for a YLIN device
using "Upper" or "Lower" as the [Matrix Format].  Note that those
format labels are case-sensitive.

It also updates one of the errorexit tests to test:

  1) an invalid label on the [Matrix Format] line.

  2) An invalid file name for the TSTONEFILE model parameter.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215