Fix Test for Upper and Lower Format for TS2 Input for YLIN
The fields in Touchstone 2 files are case-insensitive.  So, this
commit updates the tests for the optional [Matrix Format] lines
to use camel-case for full, upper and lower.  This commit also
corrects a typo in the comment lines of the tests for the Frequency
Unit on the Option line.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215