Test Use of FREQ Special Variable in AC Measures
This commit updates the existing tests for EQN and FIND-WHEN
measures for .AC analyses, for the non-step case, to test the use
of the FREQ special variable in those measures.  The correctness of
the EQN measure values in the MEASURE_AC/EquationEvalTestAC.FD.prn
file, when a measure value relies on FREQ, is also now verified.

SON Bug 1298
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1298