Add a S-param device test with data generated from EM solver
The circuit tests the extended S-parameter device with both
network parameters and short circuit currents. The data was generated
from an EM solver. The results are manually compared to the results
from an EM solver because the load is linear and it can be included
in the EM simulation. They match well. The test includes a frequency
sweep of 61 freqquency points. It is a one port EM problem with
a linear resistor at the port