Add Valgrind Testing for YLIN with Freq. Domain ISC
This commit adds the valgrind tag to the Eiger2 test.  This covers
valgrind testing for the case of a YLIN device that uses per-port
frequency-domain short-circuit currents.  Once more tests are added,
the valgrind testing for this new feature might be moved to a test that
has more than one port.  (Note: the Eiger2.cir.sh file already had the
code needed for valgrind testing.)

SON Bug 1294
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1294