Add a test of SHMOD=1 to B4SOI
No tests from the BSIM group or from the CMC test suite test the BSIM
SOI 4.6 with SHMOD=1.  I discovered an error in how I had implemented
this device that impacted ONLY SHMOD=1, and had no way to confirm that
my fixes are correct (though I know that what I did is correct and
that the code before was wrong).

Since I have NOTHING to use to verify correctness, the best I can do
is to make a test that fails before the fix and passes after.  Whether
what it does after the fix is right is not something I can test at
this point, I can only assert it.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1302
Xyce/Xyce#5