Improve Testing of 1-Point RMS and AVG Measures
This commit improves the tests for RMS and AVG measures that only
have 1 valid point in their measurement windows.  That case is now
also tested for .TRAN.  In addition, one measure from each test, has
been added to the .PRINT line to ensure that the getMeasureResult()
functions work correctly when invoked by both .MEASURE and .PRINT
statements.

SON Bugs 1267 and 1282
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1267
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1282