Test 1-Port YLIN Device with ISC_FD in Non-Linear Circuit
This commit adds a test for a 1-port YLIN device with frequency-domain
short-circuit currents.  It covers the case of a non-linear circuit.
The "original gold standards" from HSPICE are included as the files
that end in .printhr and .printhb02.  The Xyce gold standards are
modified versions of those files.  For the frequency-domain Xyce gold
standards, the files are made "two-sided" and the VM(1) value at
each non-DC frequency is then divided by two.  The VP(1) value
at a negative frequencies is the negative of the value at the
positive frequency.  Because of the differences in the scale of
the VM(1) and VP(1) values, and limitations in our AC comparator,
separate frequency-domain output files are made for the VM(1)
and VP(1) values.

SON Bug 1294
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1294