Test 2-Port YLIN Device with ISC_FD in Non-Linear Circuit
This commit adds a test for a 2-port YLIN device with frequency-domain
short-circuit currents.  It covers the case of a non-linear circuit.
The "original gold standards" from HSPICE are included as the files
that end in .printhr0 and .printhb0.  The Xyce gold standards are
modified versions of those files.  For the frequency-domain Xyce gold
standards, the files are made "two-sided".  The V-value at each
non-DC frequency is divided by 2.  The VI value at a negative
frequency is the negative of the value at the  positive frequency.
Because of the differences in the scale of the V(1) and V(2) values,
and limitations in our AC comparator, separate frequency-domain output
files are made for the V(1) and V(2) values.

This commit also cleans up the file for some other tests.  It also
moves the valgrind tag to this test, since it has more than 1-port.

SON Bug 1294
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1294