Re-do shmod=1 tests for B4SOI
The shmod=1 tests added to "test" bug 1302/issue #5 were bad, because
SHMOD=1 was totally broken.  They did test that the mistakes of that
bug changed the result, but the results they recorded were garbage.

This commit makes the shmod1 tests exactly like the equivalent shmod=0
test but with self-heating turned on (they sweep over the same
interval, use the same base model card, etc.).  That was impossible
before the fix of issue #7, because until that was fixed SHMOD=1 was
getting terrible convergence failures.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1302
Xyce/Xyce#5
Xyce/Xyce#7