Add Initial Tests for .MEASURE for .NOISE Analyses
This commit adds the initial tests for .MEASURE NOISE.  It tests
the AVG, ERR, ERR1, ERR2, EQN/PARAM, INTEG, MAX, MIN, PP and RMS
measures for the non-step case.  Because of the differences in
magnitudes, these tests only cover the V() and I() operators along
with the "control flow" for the relevant qualifiers like FROM and
TO.  Separate tests will be added for the INOISE, ONOISE, DNI() and
DNO() operators since those values are typically much smaller than
the V() or I() values.

This commit also adds the usual set of errorexit tests, related
to invalid modes and/or measure types.

SON Bug 1301
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1301