Test DERIV Measure for .MEASURE NOISE
This commit adds the initial tests for the DERIV (and DERIVATIVE)
measure for .MEASURE NOISE.  Because of the differences in
magnitudes, these tests only cover the V() and I() operators.
Separate tests will be added for the INOISE, ONOISE, DNI() and
DNO() operators since those values are typically much smaller than
the V() or I() values.

SON Bug 1301
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1301