Test INOISE, ONOISE, DNO and DNI for .MEASURE NOISE
This commit adds tests for the INOISE, ONOISE, DNO and DNI operators
for .MEASURE NOISE.  It covers the AVG, DERIV, EQN, INTEG, MAX, MIN,
PP and RMS measures.  For the DNO and DNI operators, it was deemed
necessary to test:

  1) a device (R) that has only one noise source.

  2) the total input noise from a device (Q) that has multiple
     noise sources.

  3) An individual noise source contribution (IB) from a device
     that has multiple noise sources.

The correctness of the measured values was verified manually.

SON Bug 1301
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1301