Test Touchstone 1 Input for YLIN with ISC_FD and 2-Ports
This commit adds a test for the case of a 2-port YLIN device,
where the input file is a Touchstone 1 file that also contains
per-port short-circuit current data.

It also updates the ErrorMessageTest2 test to test the case where
the algorithm for estimating the number-of-ports in the YLIN
device, from the first complete line of network data in its
Touchstone 1 input file, yields a non-integer value because the
line is not formatted properly.

SON Bug 1288
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1288