Test Compatibility of .DATA with .MEASURE DC
This commit tests the compatibility of .DATA with .MEASURE DC
for the case where .DC data=table is used in the netlist.  The FROM-TO
qualifiers now (correctly) use the table row indices as the bounds
for the measurement windows.  (Note that the table row index begins at
1 rather than 0.)  Both measure and -remeasure are tested for both
the step and non-step cases.  The ERROR measure is not tested
because it does not support the FROM and TO qualifiers.

SON Bug 1307
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1307