Test bug 812 fix
Bug 812 on Joseki-vm was about how use of .global_params to define
parameters that had names matching those of print line
accessors (which are also available in expressions and handled mostly
by the same mechanism) would break use of those accessors.

This test case defines several .global_params of that sort and then
tries to use the accessors of the same name on a print line and in
expressions on the print line.  Prior to the bug fix of commit
9e758a6c6, Xyce would exit with errors about undefined functions or
variables.  Now it works correctly.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=812