Improve FROM and TO Testing for AC, DC and NOISE Measures
This commit updates the "gold standard output" files for various
MEASURE_AC, MEASURE_DC and MEASURE_NOISE tests.  The "gold values"
printed to stdout for the start/end values for the measurement
windows should now be correct for measures that use the FROM or
TO qualifiers, as well as for the DERIV-WHEN and FIND-WHEN measures.

SON Bug 1310
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1310