Remove GetOptions setting of verbosity in test scripts
Many perl scripts were set up to call GetOptions to get the value of
any --verbose flag passed to them.  I am removing all of them.

Why?  Primarily because in commit 13bdcdee from June of 2009, the test
suite framework was changed to allow an optional extra argument to be
passed to all scripts in order to pass in the tags list.  This was
done to improve the flexibility of the VALGRIND test.  Unfortunately,
GetOptions pukes on this extra argument and so it emits a warning
about unknown options to the stderr of the script --- which is super
confusing when the test actually fails for a real reason.  This has
been confounding developers since it started, and has been an open
bug (bug 230 SON) since 2012.

Since then, the use of GetOptions to get verbosity from the script
command line has proliferated, primarily because test scripts almost
always begin life by being copied from some other.

Why remove them?  Because they're completely useless.  The top level
test suite script never passes --verbose to test scripts anyway.  It
is difficult to make it happen, and this has led to a proliferation of
scripts that explicitly set their verbosity to true irrespective of
the flag, and other scripts that blindly go through half of the
process of checking the option only to ignore it anyway.

And none of this output is of any value except to people debugging the
top-level scripts in the first place, and that can be done by forcing
verbosity on in other ways.  Debugging of lower level scripts can
always be done just by inserting debugging statements other than via
this verbosity thing --- and usually is because it's so difficult to
get --verbose working in those low-level scripts anyway.

This commit removes ALL use of GetOptions in test scripts that were
looking for verbose flags.

At the same time I've taken out commented-out calls to set verbosity,
and in some cases taken out completely wasted "use
XyceRegression::Tools;" lines in scripts that don't in fact even use
that class.

This should prevent this sort of unnecessary proliferation throughout
the test suite going forward.  Cut and paste is NOT your friend, but
at least this should prevent cutting-and-pasting of useless junk.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=230