Make xyce_verify flag nan in data as error
For as long as xyce_verify has existed, it reads in data files and
assumes they contain numbers, then goes happily along doing its
comparisons.

This is broken whenever a data file contains fields with "nan", "NaN",
or "NAN" in them.  Once these get read in, everything's nonsense, but
the code does NOT wind up computing an error norm that it detects as
excessive.

This commit changes xyce_verify so that it checks all fields that it
expects to be numbers against the string "nan" (case-insensitive
compare) and flags a data file error if it finds any.

This commit breaks the test ABM_POW/abmpow3, which in fact has a
gold-standard generation script that is broken and that creates a gold
standard with NaNs in it.

I already have a fix for test ABM_POW/abmpow3, but I'm not
committing/pushing it yet.  I want a test cycle to go through showing
that this script does in fact now flag as a failure any test that has
nan's in either its gold standard or its output, and that it does this
on every platform.  Once I see a test cycle where abmpow3 fails
everywhere, I'll commit the fix for the test.

I'm also hoping that we were not letting any nan's in output squeak
through as passes.  *shudder* That was a real possibility until this
commit, but I don't think it ever happened in regression testing.