Fix abmpow3 gold standard script
The test "ABM_POW/abmpow3" was put into the test suite in 2008, with
commit 939d98d7, as part of regression testing of joseki bugs 7 and 8,
both related to handling of negative numbers in the power operator of
the expression library.  It runs a Xyce netlist and then supposedly
compares it to a gold standard created by a perl script that should
output the same thing as Xyce.

But Xyce's ** operator does NOT really do the "right" thing when
trying to raise a negative number to a non-integer power.  Instead of
throwing a domain error, x**y returns instead sign(x)*abs(x)**y.  It
does this to avoid convergence problems that would resuit if a poorly
written ABM expression were to be fed a slightly out-of-domain value
during the course of Newton iterations.  Returning a wrong, but real
number, answer instead of "NaN" allows the newton iterations to
continue without catastrophic failure.  This is not a problem if
people construct well-behaved models, but sometimes even a
well-behaved model gets handed bad solutions during an attempt to
converge.

Perl, however, will correctly emit "nan" (or NaN, or NAN) when one
asks it to do "(-2.5)**(3.1)".  Which is exactly what the
gold-standard generation script did.

But when it was originally committed, abmpow3.cir.sh did not in fact
run abmpow3.cir.prn.gs.pl, it ran abmpow2.cir.prn.gs.pl (which is the
script for abmpow2.cir, and which tests *integer* powers).  Somehow,
the test was so insensitive that it passed, even though it was
comparing Xyce's output of, say V(1)**2.1 against V(1)**2.

8 years later, in commit 7521cc5a, this mistake was noticed and the
abmpow3.cir.sh script was changed to run abmpow3.cir.prn.gs.pl
instead.  But nobody noticed at the time that this script was throwing
NaN out the wazoo, and the script itself didn't crash.  xyce_verify
gleefully computed "NaN" as the error, but the test "$error >
tolerance" didn't trip because $error was not a number.  So it looked
like abmpow3 was passing.

A commit ago I fixed xyce_verify.pl to check data files to make
sure that they don't contain any NaNs, and abmpow3 started to fail
across all platforms, as it always should have.

This commit changes abmpow3.cir.prn.gs.pl so that it outputs what Xyce
would output (sign(x)*(abs(x)**y)) rather than x**y.  No more NaNs,
and the abmpow3 test should go back to passing.

Whether abmpow3 is a good test or not is a discussion we can have later.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=7
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=8